Refactor folder structure to prepare for code generation
