+++ Sources/BivrostBase/Export/Solidity.swift
@@ -1,0 +1,14 @@
+//
+//  Solidity.swift
+//  Bivrost
+//
+//  Created by Luis Reisewitz on 26.09.17.
+//  Copyright © 2017 Gnosis. All rights reserved.
+//
+
+/// Struct containing all Solidity Types
+public struct Solidity {}
+
+/// Struct containing all Abstract Solidity Types. Do not try to use them directly,
+/// instead use the subclasses defined in the `Solidity` struct.
+public struct _DoNotUse {}
