+++ Sources/BivrostBase/Export/SolidityContractTemplate.swift
@@ -51,13 +51,3 @@ struct StandardToken {
         }
     }
 }
-
-extension Contract.Element.Function {
-    var signature: String {
-        return "\(name)(\(inputs.map { $0.type.abiRepresentation }.joined(separator: ",")))"
-    }
-    
-    var methodId: String {
-        return String(signature.sha3(.keccak256).prefix(8))
-    }
-}
