+++ Sources/BivrostBase/Export/Types/VariableArray.swift
