+++ Sources/BivrostBase/Parser/ContractParser.swift
@@ -17,10 +17,10 @@ public struct ContractParser {
     /// - Throws: Throws if the json was malformed, e.g. a required field was missing.
     static func parseContract(from json: [String: Any]) throws -> Contract {
         guard let name = json[.contractName] as? String else {
-            throw BivrostError.Parser.contractNameInvalid
+            throw ParsingError.contractNameInvalid
         }
         guard let elementsJson = json[.abi] as? [[String: Any]] else {
-            throw BivrostError.Parser.contractAbiInvalid
+            throw ParsingError.contractAbiInvalid
         }
         
         let elements = try elementsJson.map { try ElementJsonParser.parseContractElement(from: $0) }
